requirements.txt
tensorflow==2.15.0
numpy==1.24.3
pillow==10.1.0
streamlit==1.29.0
matplotlib==3.8.2